refactor: Add cooldown data structure and functions
This commit refactors the code to add a new cooldown data structure and related functions. The new COOLDOWN object stores cooldown information for different types of abilities, such as personal, raid, and target. The functions `getCooldown` and `orderBy` have been added to retrieve and sort the cooldowns based on specified criteria.